:alien: Migrate rand from 0.8.5 to 0.9.0 and rand_chacha from 0.3.1 to 0.9.0

+++ cli/src/command/update.rs
@@ -197,7 +197,7 @@ fn update_archive<Strategy: TransformStrategy>(args: UpdateCommand) -> io::Resul
 
     let (tx, rx) = std::sync::mpsc::channel();
 
-    let random = rand::random::<usize>();
+    let random = rand::random::<u64>();
     let temp_dir_path = temp_dir().unwrap_or_else(|| {
         archive_path
             .parent()
